Refactor templatetags into two separate modules.
