+++ opal/templates/base.html
@@ -1,6 +1,6 @@
 {% load staticfiles %}
 {% load compress %}
-{% load opalplugins %}
+{% load application %}
 <!doctype html>
 <html>
 	<head>
