+++ opal/templates/navbar.html
@@ -1,5 +1,7 @@
-{% load opalplugins %}
 {% load static %}
+
+{% load application %}
+{% load plugins %}
 <nav class="navbar navbar-default navbar-primary navbar-fixed-top screen-only top-navbar" role="navigation">
   <div class="container-fluid">
     <!-- Brand and toggle get grouped for better mobile display -->
