+++ opal/templates/opal.html
@@ -1,7 +1,8 @@
 {% load staticfiles %}
 {% load compress %}
 
-{% load opalplugins %}
+{% load application %}
+{% load plugins %}
 {% load gifs %}
 <!doctype html>
 {% block angulardefinition %}
