+++ opal/templatetags/plugins.py
@@ -1,11 +1,11 @@
 """
-Templatetags for including OPAL plugins
+Templatetags for working with OPAL plugins
 """
 import itertools
 
 from django import template
 
-from opal.core import application, plugins
+from opal.core import plugins, application
 
 register = template.Library()
 
@@ -36,7 +36,6 @@ def plugin_head_extra(context):
     ctx['head_extra'] = templates
     return ctx
 
-
 def sort_menu_items(items):
     # sorting of menu item is done withan index property (lower = first), if they
     # don't have an index or if there are multiple with the same
@@ -46,7 +45,6 @@ def sort_menu_items(items):
     index_sorting = lambda x: x.get("index", 100)
     return sorted(sorted(items, key=alphabetic), key=index_sorting)
 
-
 @register.inclusion_tag('plugins/menuitems.html')
 def plugin_menuitems():
     def items():
@@ -55,15 +53,6 @@ def plugin_menuitems():
                 yield i
 
     return dict(items=sort_menu_items(items()))
-
-
-@register.inclusion_tag('plugins/menuitems.html')
-def application_menuitems():
-    def items():
-        app = application.get_app()
-        for i in app.menuitems:
-            yield i
-    return dict(items=items)
 
 @register.inclusion_tag('plugins/angular_module_deps.html')
 def plugin_opal_angular_deps():
@@ -88,38 +77,3 @@ def plugin_opal_angular_tracking_exclude():
         excluded_tracking_prefix=yield_property("opal_angular_exclude_tracking_prefix"),
         excluded_tracking_qs=yield_property("opal_angular_exclude_tracking_qs")
     )
-
-@register.inclusion_tag('plugins/javascripts.html')
-def core_javascripts(namespace):
-    def scripts():
-        app = application.get_app()
-        for javascript in app.core_javascripts[namespace]:
-            yield javascript
-    return dict(javascripts=scripts)
-
-@register.inclusion_tag('plugins/javascripts.html')
-def application_javascripts():
-    def scripts():
-        app = application.get_app()
-        for javascript in app.javascripts:
-            yield javascript
-    return dict(javascripts=scripts)
-
-@register.inclusion_tag('plugins/stylesheets.html')
-def application_stylesheets():
-    def styles():
-        app = application.get_app()
-        for style in app.styles:
-            yield style
-    return dict(styles=styles)
-
-@register.inclusion_tag('plugins/actions.html')
-def application_actions():
-    def actions():
-        app = application.get_app()
-        for action in app.actions:
-            yield action
-        for plugin in plugins.plugins():
-            for action in plugin.actions:
-                yield action
-    return dict(actions=actions)
